fix(mitm-addon): break model pricing import cycle
