+++ crates/runner/mitm-addon/src/model_usage_pricing.py
@@ -1,4 +1,4 @@
-"""Verification for signed model token price schedules."""
+"""Signed response-header verification for model token price schedules."""
 
 import base64
 import binascii
@@ -6,28 +6,18 @@ import hashlib
 import hmac
 import json
 import time
-from collections.abc import Mapping
-from dataclasses import dataclass
 
 from mitmproxy import http
 
 import flow_metadata
 import flow_metadata_keys as metadata_keys
-from usage.model_tokens import MODEL_USAGE_CATEGORIES
+from usage.model_pricing import ModelUsagePricing, parse_model_usage_pricing
 
 PRICING_HEADER = "x-vm0-usage-pricing"
 PRICING_SIGNATURE_HEADER = "x-vm0-usage-pricing-signature"
 _PRICING_SIGNATURE_DOMAIN = b"vm0-model-usage-pricing-v1\0"
 _MAX_PRICING_BYTES = 2048
 _MAX_CLOCK_SKEW_SECONDS = 300
-
-
-@dataclass(frozen=True)
-class ModelUsagePricing:
-    """Validated model token prices used while constructing usage events."""
-
-    unit_size: int
-    unit_prices: dict[str, int]
 
 
 def apply_signed_usage_pricing(flow: http.HTTPFlow) -> bool:
@@ -84,14 +74,6 @@ def apply_signed_usage_pricing(flow: http.HTTPFlow) -> bool:
     return True
 
 
-def from_flow_metadata(meta: Mapping[str, object]) -> ModelUsagePricing | None:
-    """Read one complete pricing schedule from primitive flow metadata."""
-    value = meta.get(metadata_keys.MODEL_USAGE_PRICING)
-    if not isinstance(value, dict) or set(value) != {"unitSize", "unitPrices"}:
-        return None
-    return _pricing_from_fields(value["unitSize"], value["unitPrices"])
-
-
 def _bearer_token(value: str) -> str | None:
     if not value.startswith("Bearer "):
         return None
@@ -139,19 +121,4 @@ def _decode_pricing(value: str) -> ModelUsagePricing | None:
     if abs(int(time.time()) - issued_at) > _MAX_CLOCK_SKEW_SECONDS:
         return None
 
-    return _pricing_from_fields(pricing["unitSize"], pricing["unitPrices"])
-
-
-def _pricing_from_fields(unit_size: object, raw_unit_prices: object) -> ModelUsagePricing | None:
-    """Construct an atomic schedule or reject all of its prices."""
-    if not isinstance(unit_size, int) or isinstance(unit_size, bool) or unit_size <= 0:
-        return None
-    if not isinstance(raw_unit_prices, dict) or set(raw_unit_prices) != set(MODEL_USAGE_CATEGORIES):
-        return None
-    unit_prices: dict[str, int] = {}
-    for category in MODEL_USAGE_CATEGORIES:
-        unit_price = raw_unit_prices[category]
-        if not isinstance(unit_price, int) or isinstance(unit_price, bool) or unit_price < 0:
-            return None
-        unit_prices[category] = unit_price
-    return ModelUsagePricing(unit_size=unit_size, unit_prices=unit_prices)
+    return parse_model_usage_pricing(pricing["unitSize"], pricing["unitPrices"])
